+++ R/dummy_functions.R
@@ -78,10 +78,7 @@ example_datasets <- function() { # nolint
 #' @return A `teal` module which can be included in the `modules` argument to [teal::init()].
 #' @examples
 #' app <- init(
-#'   data = teal_data(
-#'     dataset("IRIS", iris),
-#'     dataset("MTCARS", mtcars)
-#'   ),
+#'   data = teal_data(IRIS = iris, MTCARS = mtcars),
 #'   modules = example_module()
 #' )
 #' if (interactive()) {
@@ -93,22 +90,27 @@ example_module <- function(label = "example teal module", datanames = "all") {
   module(
     label,
     server = function(id, data) {
-      checkmate::assert_class(data, "tdata")
+      checkmate::assert_class(data(), "teal_data")
       moduleServer(id, function(input, output, session) {
-        output$text <- renderPrint(data[[input$dataname]]())
+        ns <- session$ns
+        updateSelectInput(session, "dataname", choices = isolate(teal.data::datanames(data())))
+        output$text <- renderPrint({
+          req(input$dataname)
+          data()[[input$dataname]]
+        })
         teal.widgets::verbatim_popup_srv(
           id = "rcode",
-          verbatim_content = attr(data, "code")(),
+          verbatim_content = reactive(teal.code::get_code(data())),
           title = "Association Plot"
         )
       })
     },
-    ui = function(id, data) {
+    ui = function(id) {
       ns <- NS(id)
       teal.widgets::standard_layout(
         output = verbatimTextOutput(ns("text")),
         encoding = div(
-          selectInput(ns("dataname"), "Choose a dataset", choices = names(data)),
+          selectInput(ns("dataname"), "Choose a dataset", choices = NULL),
           teal.widgets::verbatim_popup_ui(ns("rcode"), "Show R code")
         )
       )